ci: amplify push
