+++ amplify/.config/project-config.json
@@ -1,4 +1,7 @@
 {
+  "providers": [
+    "awscloudformation"
+  ],
   "projectName": "pretraining",
   "version": "3.1",
   "frontend": "javascript",
@@ -10,8 +13,5 @@
       "BuildCommand": "yarn build",
       "StartCommand": "yarn start"
     }
-  },
-  "providers": [
-    "awscloudformation"
-  ]
+  }
 }
