+++ amplify/backend/function/wandbHistory/Pipfile
@@ -1,0 +1,15 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+
+[packages]
+src = {editable = true, path = "./src"}
+boto3 = "==1.29.5"
+wandb = "==0.16.0"
+pandas = "==2.0..3"
+
+[requires]
+python_version = "3.8"
