+++ amplify/backend/function/wandbHistory/src/src.egg-info/PKG-INFO
