+++ amplify/team-provider-info.json
@@ -21,7 +21,7 @@
         "wandbHistory": {
           "wandbApiKey": "d99387d278a657c2e12313a7758ac33f3de8fd26",
           "deploymentBucketName": "amplify-pretraining-prod-115707-deployment",
-          "s3Key": "amplify-builds/wandbHistory-5a4672446f2f2f36414b-build.zip"
+          "s3Key": "amplify-builds/wandbHistory-5a58544a6f506737304d-build.zip"
         }
       },
       "auth": {
